show list student of room and pick leader of room and optimize code
