+++ ptit_dormitory_be/routes/areaRoute.js
@@ -7,16 +7,16 @@ areaRoutes.post('/createPlaces', areaController.createPlaces);
 areaRoutes.put('/updatePlaces/:id', areaController.updatePlaces);
 
 // Test
-areaRoutes.get('/area', areaController.getAreas);
-areaRoutes.post('/area', areaController.addNewArea);
-areaRoutes.put('/area/:id', areaController.updateArea);
-areaRoutes.get('/floor', areaController.getFloorsByArea);
-areaRoutes.post('/floor', areaController.addNewFloor);
-areaRoutes.put('/floor/:id', areaController.updateFloor);
-areaRoutes.get('/room/getRoomsByFloor', areaController.getRoomsByFloor);
-areaRoutes.get('/room/getRoomsByArea', areaController.getRoomsByArea);
-areaRoutes.get('/room', areaController.getAndFindRooms);
-areaRoutes.post('/room', areaController.addNewRoom);
-areaRoutes.put('/room/:id', areaController.updateRoom);
+// areaRoutes.get('/area', areaController.getAreas);
+// areaRoutes.post('/area', areaController.addNewArea);
+// areaRoutes.put('/area/:id', areaController.updateArea);
+// areaRoutes.get('/floor', areaController.getFloorsByArea);
+// areaRoutes.post('/floor', areaController.addNewFloor);
+// areaRoutes.put('/floor/:id', areaController.updateFloor);
+// areaRoutes.get('/room/getRoomsByFloor', areaController.getRoomsByFloor);
+// areaRoutes.get('/room/getRoomsByArea', areaController.getRoomsByArea);
+// areaRoutes.get('/room', areaController.getAndFindRooms);
+// areaRoutes.post('/room', areaController.addNewRoom);
+// areaRoutes.put('/room/:id', areaController.updateRoom);
 
 export default areaRoutes;
